requirements.txt
Django>=2.1.3,<2.2.3
djangorestframework>=3.9.0,<3.10.2
psycopg2>=2.7.5,<2.8.0
requests<=2.22.0
flake8>=3.6.0,<3.7.8
pytest>=5.0.1
mock==3.0.5
